requirements.txt
pandas
openpyxl
requests
gspread
google-auth
streamlit
google-generativeai
matplotlib